Fix handling initializers for structs with anonymous scalar members
In r32265 / cb57ee43 the handling of initializers for structs
was changed in order to fix the handling of unions:
Before that, anonymous union members were wrongly ignored,
the old code was basically doing
"!(isAnonymous && !isArrayOrStruct)"
and filtered all anonymous scalar and union members out.
That commit changed it to
"!(isAnonymous && !isArrayOrStruct && !isUnion)"
but led to crashes for scalar members.
The fix for the crash from r32292 / 98173c24 accidentally changed it to
"!(isAnonymous && !isArrayOrStruct && isElaborated && !isUnion)"
but correct would be
"!(isAnonymous && !isArrayOrStruct && !(isElaborated && isUnion))".
Due to this all scalar anonymous members were wrongly kept
instead of being filtered out.

Now we use a utility method to make this easier to understand,
and actually the check for "is array or struct or union"
is also easier if we ignore the isAggregateType utility method.

Fixes #661

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@40839 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c