Add limited check for happens-before relationship.
Operations issued by the same thread never create a conflict,
and memory accesses of a thread also do not race with accesses
of the parent thread made before the thread was created
(at least that seems to be the assumed behavior for SV-COMP).
Depending on the assumed memory model and supported operations
we could rule out more potential races.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/data-race-detection@40849 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c