Improve yaml witness parsing into POJOs
There are different entry types, so we need to reflect this cleanly
when reading in the yaml witnesses. So far only the loop_invariant
entry type was supported. This commit adds support for
loop_invariant_certificate and restructures the code such that new
entry types can be easily added in the future. We now call the
different fields of an entry (apart from entry_type) "records".
Many of these records will be used by entries of different
entry types, so we move them to a new package *.records.common.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@40864 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c