Add auto-indentation feature to Rewrite class
This is very simple. We just try to determine how many whitespace
is preceding the position where we insert some code blob.
We then indent all following lines of the inserted snippet by
that prefix. So this would even work if the line has something
like " \t  ", we would then just prepend this to all following
lines of the snippet.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/loopsummary@40903 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c