Replace pairs of irrelevant AssumeEdges in SliceToCfaConversion
The goal is that the CFA does not contain AssumeEdges with
expressions with undefined variables, even when the CFASimplifier
is not used, so that the resulting export program is compileable.

Disabling the use of the CFASimplifier is necessary for the
evaluation of #994 as the changes made by the CFASimplifier are
not recorded.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/994-CFA-to-C-exporter-with-CFA-transformations@40919 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c