Add an option to ARGCPA about what to do on late merges
A "late merge" is a merge of two abstract states
where one of the two abstract states already has at least one child state.
Performing a merge in such a situation may be undesired for some
configurations. For other configurations, it may be a bug
that such a situation even happens.

So we add an option that allows us to configure what happens.
ALLOW is what we previously had as the only option,
and what remains the default.
PREVENT is for configurations that do not want late merges
but also do not want to use merge-sep because they are fine with
non-late merges.
CRASH and *_WARN are debugging utilities
for cases where late merges should never happen.

Part of #991

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@40970 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c