Add cpa.arg.lateMerge=prevent to predicate analysis for cex checks
A discussion for this is in #716 and #907,
and late merges are explained in #991.

A first attempt of avoiding late merges was made by changing
the traversal order in r40787 / 9abc094a,
but this had to be reverted in r40843 / 44aaf4bb due to #990.

As a workaround until we have a better traversal strategy for
counterexample checks and witness validation,
we can at least avoid late merges in the ARGCPA.
This will not guarantee that at most one SAT check is done during
counterexample check, but at least those checks won't be redundant,
and the unwanted ARG structure with branchings will not occur.
The latter fixes crashes for -predicateAnalysis-slicing.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@40973 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c