Add cpa.arg.lateMerge=prevent to BMC and all BMC-based configs
This is a first step (if not the solution) for #1002.
Late merges are explained in #991.

With this change, crashes due to unexpected ARG branching
during counterexample analysis of BMC
(and thus all BMC-based analyses like k-induction, IMC, etc.)
should be gone.
However, the ARG will still be larger than if the traversal order
would not produce such situations in the first place.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@40978 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c