Fix missing aliasing constraints for local variables in predicate analysis
This would happen if a local variable of some function is used
only on some branches of the code, but then later on the function is
called again also on other branches, and if between these calls the
abstract states were merged.
The reason for this is that aliasing constraints were only added
once the variable was seen for the first time,
but only for those paths where it was seen first.
After merging abstract states the variable was considered to be known
already for all paths, though, even for those paths on which the
variable was not actually seen and no constraints had been created.

This affected only local variables which had their address taken
or for which aliasing constraints are relevant for other reasons.

The fix is to make sure when merging abstract states
that the required aliasing constraints are present for all paths.

Fixes #987

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@40994 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c