Fix crucial bug in naive loop acceleration
Without this assumption, there exists a bypass path in the CFA
via which this abstract acceleration strategy is essentially the
same as the havoc strategy. Only with this assumption we ensure
that the nondeterministic values that we havoc are constrained
to those values that actually lead to entering the loop.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/loopsummary@41067 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c