Remove hacky code from LocationTransferRelation
Setting the SummaryInformation like this is not a good idea.
There are analyses like BMC that run their own sub-analysis
like the TargetLocationProvider. Then this would mess up things
like the strategy selection strategy etc. On the long run,
we should adhere to using precision instead of a mutual god object
that is attached to the CFA.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/loopsummary@41072 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c