Refactor code to be less error-prone
We do not want to declare the next "new" CFA node in advance every
time. errorprone complains about the last instance where we do not
use this new value, and rightfully so. Ideally, a new node is always
just created at the point where it is really needed. The current
node can be updated by returning it from the method that implemented
the change on the CFA. This way we are less likely to screw up
by forgetting to update the references to the current CFA nodes.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/loopsummary@41078 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c