update JavaSMT from v3.11.0 to v3.12.0
API change applied:
- adding size-method in ProverEnvironment.
- support String-theory in Model.

Please note that for delegating ProverEnvironments,
we assume a simple delegate-method for the size-method.
If the delegated ProverEnvironment has already some size
before being nested in a delegating ProverEnvironment or
is modified in any other way except through the wrapped itself,
the returned size will include the additional operations.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/1007-update-JavaSMT-from-v3.10.1-to-v3.13.0@41114 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c