Fix blank edge chain detection in `CFASimplifier`
In the old version, `findEndOfBlankEdgeChain` doesn't consider
the number of entering edges of a node and could go past the
merge point of a branching.
It doesn't seem like this was the author's intent, even though
the old code worked, because `removeChainOfNodes` does
consider the number of entering edges. Also, the code for
removing outgoing blank edges from a merge point wouldn't
actually do anything.
This commit fixes the behavior of `findEndOfBlankEdgeChain`
and make the rest of the code easier to understand.
The overall behavior of `CFASimplifier` should stay the same.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@41185 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c