Fix blank edge chain detection in `CFASimplifier`
In the old version, `findEndOfBlankEdgeChain` doesn't consider
the number of entering edges of a node and could go past the
merge point of a branching.
It doesn't seem like this was the author's intent, even though
the old code worked, because `removeChainOfNodes` does
consider the number of entering edges. Also, the code for
removing outgoing blank edges from a merge point wouldn't
actually do anything.
This commit fixes the behavior of `findEndOfBlankEdgeChain`
and makes the rest of the code easier to understand.
The overall behavior of `CFASimplifier` should stay the same.

Commit r41185 was supposed to fix this, but the commit lead to
assertions, because the merge point can have other entering
edges that prevent removal of the merge point. This is why
r41185 was reverted by r41201.
This commit fixes what r41185 was supposed to fix and also
takes additional entering edges of the merge point into account.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@41295 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c