Insert blank edge to mark block ends
We mark block ends with a blank edge. We need this edge to strengthen the analysis in the very last step with the error conditions from the successors. In case the block end is reachable with the strengthening, we have to perform a backward analysis. Otherwise, we know that the error is not reachable and will not ever be reachable from this block. We proceed to broadcast the abstraction to exclude this infeasible error path from further analysis.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/abstraction-with-block-summaries@41594 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c