Make test cases less specific to implementation
This enables that the tests will pass even if someone changes the
formatting/whitespace of the formulas that the
FormulaToCExpressionConverter generates. Also if there are other
changes in the future like with parentheses, this change now allows
us to update the testing logic without having to change every test
method by collecting the logic of what property we test in each case
in a new method that is called from every test method.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/cinvariants-export@41706 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c