Add method CType.hasKnownConstantSize()
Whether a type has a statically known size
is actually different from whether the type is complete,
but so far we use only the latter.
In order to properly handle variable-length arrays and avoid #1031
we need to distinguish them.
So as a first step towards #1031 we add this information to CType.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@42204 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c