Clean up benchmark definitions for SV-COMP benchmarks
- There was no benchmark definition for -svcomp23 yet,
  making local pre-runs and experiments more difficult than necessary.
- There were lots of useless old benchmark definitions lying around,
  which cannot be used anymore due to changes in *.set files etc.

In contrast to our old -svcomp* configs, which we keep as unmaintained
on purpose such that users can for example still call CPAchecker
with -svcomp22, there is no use case for old benchmark definitions,
because these are just for CPAchecker's developers and do not influence
CPAchecker's behavior.
They cannot even be used to reliably look up what options where used
back then because nobody knows whether they have been consistently kept
in sync with the official benchmark definition in the SV-COMP repo
(likely not).
So we can just delete them and anybody who really needs them can look
in the repository history.

In order to make it easier to keep an up-to-date benchmark definition
that copies the real SV-COMP one,
lets avoid the set of individual benchmark definitions for each
property and keep just one file.
Any change to the official benchmark definition can likely be applied
directly to this copy here.
When executing the benchmarks, one can (and must) still execute only
a single run definition, which can be done with the "-r" parameter to
benchexec.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@42212 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c