Fix calculation of array sizes limited to ints in ASTConverter
If there is no explicit length in the array declarator,
we calculate the length from the initializer.
This logic used "int" so far, which could in theory be too small
for huge arrays.
Because we will need a BigInteger in the end anyway,
we can easily fix this by performing the whole calculation with
BigInteger.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@42228 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c