Simplify handling of enum constants in frontend
After recursively simplifying an expression,
we will not have unary expressions for negative numbers etc.,
so we can simplify this handling.
This was likely necessary only a long time ago in the past.

We now produce a parse error instead of just a warning
if there is something unexpected in the value of an enum constant.

And there is now a utility method for constant expressions,
which can likely be reused in similar places.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@42229 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c