Fix crash for array designators that are not integer literals
In the C code "int a[] = { [foo] = 1 };"
the foo is an integer constant expression.
We can handle this, but our frontend blindly assumed that it is a
CIntegerLiteralExpression, which it typically is because we simplify
constant expressions as far as possible.
However, with "cfa.simplifyConstExpressions=false" or in cases where not
a legal constant expression is used, it would crash with a
ClassCastException.
Now we handle this even if "cfa.simplifyConstExpressions=false"
and produce a proper parser error otherwise.

Note that this does not simplify the produced AST,
as only the internal calculation of the array length was affected.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@42230 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c