Better handling of variable-length arrays in predicate analysis
This only replaces a few calls to CType#isIncomplete()
with !CType#hasKnownConstantSize()
such that the fallbacks for incomplete types
are also applied for variable-length arrays.
This is not a proper fix,
but the fallbacks for incomplete types
(which typically rely on config options like defaultArrayLength)
are at least better than the wrong results returned by getSizeof().

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@42235 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c