Add test tasks for tracking buggy type calculation
For sizeof and _Alignof, we currently interpret the types as int
instead of size_t aka unsigned long int, which can be showcased by
the tasks added in this commit. The bug is visible both in reachability
as well as in overflows. The tasks are TRUE with respect to those
labels, but CPAchecker currently reports them as FALSE, e.g. when
checking with predicateAnalysis or predicateAnalysis--overflow.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@42283 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c