Extend test tasks to cover sizeof(x) and alike
The previous tasks where just for type id expressions, i.e., when
sizeof or _Alignof are called with a type name. But there is also
the other case where we pass things like variable names, which will
be handled as unary expression instead. These tests make sure we get
the type correctly also in these cases, which it seems like we already
do, so there is nothing to fix. It is still good to have these
analogous tests for finding regressions in case we ever change the
logic for these special operators.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@42288 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c