Fix types for _Alignof again, this time for real (:fingers_crossed:)
cf. #237, r19243 / 10692b2f, r20582 / c43673c1, and r42284 / 51f8bfc0

In the past, we had likely misunderstood the sentence "The result is
an integer" in the C standard with regards to _Alignof to mean that the
result has type int, so we hard-coded that in order to override the
different type that CDT gave us (a pointer type).
But this is wrong, the standard actually says "its type is size_t",
so yesterday the old hard-coded int was reverted.
But it turns out that CDT still gives a (different) wrong type for
_Alignof, namely int.
So let's go back to hard-coding the type, this time correctly with
"unsigned long int", which should be the same as size_t on 32-bit and
64-bit.

For _Alignof on expressions instead of types,
we still had the wrong hard-coded int, this was not reverted yesterday.
But now it is also fixed.

In order to take no chances with CDT, we just do the same for sizeof.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@42303 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c