Use long instead of int for type sizes in PredicateCPA
This is a partial fix for #1038.
Using long still does not cover all potential sizes,
but is at least much better than int.
We also do not fix all cases where type sizes are used,
because for example we cannot create bitvectors in JavaSMT
with a length > MAX_INT.
But it at least prevents a false alarm from the MWE in #1038
(and turns it into a timeout due to #1039).

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@42308 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c