Fix handling of __VERIFIER_atomic with local lock
Per SV-COMP rules, functions starting with __VERIFIER_atomic
are important for threading, since they are guaranteed to be executed
atomically. We already set a global access lock for them, but did not
yet mark them as important regarding the local access lock. That way,
it was possible to overlook certain interleavings, which lead to a
unsound analysis.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@42344 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c