Copy concurrency options between configs
We already have a -predicateAnalysis-concurrency, so I copy
some of the options that seem plausible over to:
-predicateAnalysis-concurrency--overflow
My guess is that these options should also work fine there as well.
I ran this on ConcurrencySafety-NoOverflows and the score is
the same, but this give about 15% speedup, so we will take it.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@42381 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c