Fix broken handling of integer promotions for unary ops in PredicateCPA
If we apply integer promotion before the operation,
we need to use the promotion type as the starting point
when casting the operation's result.
This affects unary minus and bitwise negation
when applied on types smaller than int,
but only if the resulting expression type is also smaller than int.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@42391 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c