Mark pthread_cond_timedwait as supported.
The only difference to pthread_cond_wait is that whenever pthread_cond_timedwait returns,
the return value could indicate a timeout.
Since the ThreadingCPA currently does not implement special handling for other error-indicating
return codes (except for pthread_mutex_trylock where this is crucial) it is probably fine to
ignore this here for now as well.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/threading-extensions@42421 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c