Disable serialization warnings from javac
On Java 18 there are a lot more of them,
and many are irrelevant.
There are often classes that are promised to serializable
if some user-supplied content is serializable (e.g., data structures),
and the compiler warns in such cases.
We would have to add over 100 suppressions,
so lets disable the warning instead.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@42437 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c