Make hash code of symbolic-expression classes deterministic
Fixes #1024

The hashCode methods of these classes use the hash codes of Class
objects, which is correct, but which leads to hashes that can change
from one run of CPAchecker to the next.
And this can make results non-deterministic.

Of course, relying on HashMaps is not a good choice when one wants
deterministic output, but in this case there does not seem to be a
downside in making the hashCode consistent over CPAchecker runs as well.

Thanks to @canliture for bringing this up and suggesting the solution.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@42470 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c