Use ImmutableSortedMap in TransitionCondition.Scope for usedDeclarations
This does not actually change the order of the map, because all existing
callers of the constructor already passed a sorted map, and
ImmutableMap.copyOf() keeps the order.
But the implementation of compareTo() relies on the map being sorted
consistently, because otherwise lexicographically comparing the entries
does not make sense.
So it is better to actually enforce and document the sortedness of this
map.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@42473 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c