Turn off tracking "fields" in PointerTargetSet for array-based encoding
As discussed in
https://gitlab.com/sosy-lab/software/cpachecker/-/issues/1039#note_1177874407
the information about fields is wrong without the tracking of targets.
However, the information about fields is used only in one place,
and there does not seem to be any performance advantage,
at least in combination with the array-based encoding.

So we can just disable the tracking of "fields" completely
and let "tracksField()" always return "true".

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/1039-optimize-predicate-target-tracking@42477 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c