Make handling of deferred allocations more consistent in PredicateCPA
All code except for the one changed here that handles deferred
allocations is guarded by the options deferUntypedAllocations
and revealAllocationTypeFromLhs.
This does not actually matter because if both options are false,
no deferred allocations are created and the respective code will do
nothing.
But in other similar places the option is also checked,
so lets do it here as well for consistency and clarity.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@42510 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c