Refactoring regarding base-address constraints in PointerTargetSetBuilder
Right now, addBase() and prepareBase() are responsible for adding the
base-address constraints, so callers do not have to care about this.
However, this gets inconvenient, for example for fixing #912
where we need to add the constraints at a different time than the base.

They also need different arguments.
Some parameters are needed only for the base-address constraints.
On the other hand, the type is required for addBase() and prepareBase(),
but we want to make it optional for the base-address constraints
for being able to fix #912,
which does not really work if the callers require it.

There are only a few internal callers, so lets put the responsibility on
them and move the call for creating the base-address constraints
out of addBase() and prepareBase().
And then we can mark the type argument as optional,
because it is not really required for allocations of dynamic memory.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@42511 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c