Do not defer base-address constraints for deferred allocations
Fixes #912

If there is an allocation with an unknown type
("void* p = malloc(x);"), we defer handling the allocation until at some
later point hopefully we see which type the program uses to access the
allocation.
Part of the handling of an allocation is adding constraints
like the one that ensures that allocations do not overlap.

Now, if the type of an allocation was never revealed,
these constraints were missing, which is imprecise.

But actually, we can add these constraints immediately,
because they do not depend on the type
(for such allocations we always have the argument of malloc etc.
that tells us the correct size independently of the LHS type).
So lets do this.

We just need to make sure that we do not add the constraints twice,
because this would create conflicting constraints and be unsound.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@42514 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c