Mark Refaster CI job as allowed to fail
Somehow the NullPointerException that we previously only saw with
Error Prone 2.16 (https://github.com/google/error-prone/issues/3484)
has now started to appear with Error Prone 2.15 as well.
We will wait for Error Prone 2.17 and then try again.

This is not critical as Refaster suggestions come with a patch,
so even if lots of issues are introduced in the meantime,
we can easily fix them all automatically later on.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@42571 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c