Use pattern matching for instanceof
We can now add a variable identifier to an instanceof expression,
and this declares a new variable with the value and type of the
instanceof expression.
This even works if the instanceof expression is negated,
but whether such cases are good to read is unclear.
This change applies the refactoring only to non-negated expressions.

Commit automatically generated by Eclipse.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/993-java-17@42724 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c