Refactoring: merge if conditions using instanceof pattern matching
Previously there was often code like

if (a instanceof Foo) {
  Foo f = (Foo) a;
  if (f.bar()) {

This can now be rewritten into

if (a instanceof Foo f && f.bar()) {

The first part was done in the previous commit,
this commit merges the two if conditions.
The diff is big due to indentation changes,
but fine when using --ignore-space-change.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/993-java-17@42725 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c