Fix if condition in ExpressionSimplificationVisitor that was never true
This prevented inlining initializers of const variables
into uses of these variables as part of our CFA simplification.
Such inlining was obviously desired because the code was written,
but not working at least for several years.
If undesired, it can be disabled with cfa.simplifyConstExpressions=false.

Found by marking the type hierarchy as sealed in r42750 / b443f28e.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@42753 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c