Use a record instead of Triple in SMGPathInterpolator
This is an interesting example because the Triple was only used locally
inside one function, not as part of some API,
so it wasn't that bad.
But it contained a boolean and an int for which the semantics were not
obvious.
Declaring a record is so little effort and even possible within a
function, so it is worth it here.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/993-java-17@42764 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c