Refactor test utility: remove unused objects
The utility method returned a triple with two edges and a CFA object,
but actually only the first edge was ever used.
But additionally, the method also had a side effect of assigning yet
another edge to a field in the class.
To clean this up, we remove the return value
and just create and store the two actually used edges.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@42779 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c