Reference line offsets by hash instead of filename.
The current behavior is problematic because invariants are skipped
by the parser if the path to the program file stored in the witness
is different from the path given by the config used to parse the witness.
This commit removes the need to recreate the directory structure
from when the witness was created.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/vercors-annotations@42958 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c