Make add/remove methods in `MutableCFA` behave like add/remove of collections
The behavior of `MutableCFA#addNode(CFANode)` and `MutableCFA#removeNode(CFANode)`
was undocumented and it wasn't really clear what calling these methods entails.
This commit changes those methods to behave like add/remove methods of collections,
which is likely what users expect, and adds documentation to the the methods
that describe the new behavior.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/cfa-network@42964 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c