Fix infinite loop in AssumptionToEdgeAllocator for variable-length arrays
AssumptionToEdgeAllocator relies on a statically known array size
for its bounds check, and if this size is not known,
it does no bounds check and loops infinitely.
Before r42979 / b3feab3b this did not occur because for VLAs
a wrong array size was used (cf. #1031)
and only for incomplete array types the bounds check was skipped,
but for incomplete array types we had a separate check anyway.

Since that commit, we correctly avoid using the wrong array size,
but this meant that now there was no bounds check at all for VLAs.

Now we change that check that for all arrays without known size
we just stop immediately and for other arrays we do the bounds check.
This also covers incomplete array types
because isIncomplete() implies !hasKnownConstantSize() for arrays.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@42981 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c