Fix execution of Black in CI and pin major version of Black
r42210 / b2fbe0ced12 attempted to let Black run additionally on contrib/,
but made it such that it does not run on our other scripts anymore.
Furthermore, it only changed the CI config,
which is not useful for people running Black locally.

So now we do this properly.

Furthermore, each year Black might have some formatting changes.
In order to not have failing CI runs on all branches then,
lets pin it to the current major version.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@43021 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c