Avoid multi-variable declarations (complex cases)
These are cases where at least one of the declared variables
is actually only required to have a smaller scope
than its current declaration.
This makes the code easier to understand because often the new scope
is significantly smaller or within a loop
(such that the declaration now makes it clear that the variable
is not reused across loop iterations).

When looking at this, note that it is safe to move a declaration
into a smaller scope if the variable is initialized at the new
declaration. This means that any value it would previously have from a
previous loop iteration would have been overwritten anyway.
This is the case in all the changes in this commit.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@43126 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c