Make temporary variables for casts of field accesses optional
Additional temporary variables for pointer accesses have been introduced
as part of #110 and for operands of casts this was r8328 / 8f3497a8.
Unfortunately, in this commit it was forgotten to hide this behind
the cfa.simplifyPointerExpressions option from r8153 / 1a2b460c.
Let's do that now before the 10th anniversary of that change. :-)

To be clear, previously for C code like "(void *)test->member"
a temporary variable was added, now it isn't anymore
(at least by default).

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@43209 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c