Add support for memcpy/memmove non-literal sizes and fractional elements
The support is realized by a new slice assignment method in
AssignmentHandler, which can assign to multiple elements at will. To
make this work, it is needed to make it possible to write array elements
conditionally, so a method was added to SMTHeap so that it is possible
to assign old-indexed array to new-indexed one without changing any
elements. Since I am unsure how it should be implemented in
SMTHeapWithUninterpretedFunctionCalls, I did not implement it yet, it
just throws for now.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/predicate-memory-functions@43218 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c