Add memset non-literal sizes by introducing slice expresssions
For memset, we need to be able to recursively set all structure fields,
which may result in nested indexing and field accesses, for example,
foo[i].a[j].b.c[k], where i is the memset destination index and j,k are
indices for arrays in structures, which should be set fully. As such, I
introduced ArraySliceExpression, which is able to store these
expressions, unlike CExpression.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/predicate-memory-functions@43219 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c