Move type field of CEnumerator into CEnumType
All enumerators of an enum share the same type,
and furthermore this type can only be determined
until after all CEnumerator instances are created.
So it makes more sense to store it in CEnumType.
Also cf.
https://gitlab.com/sosy-lab/software/cpachecker/-/issues/348#note_193322139

In order to make this possible,
we need to move the type field out of AbstractSimpleDeclaration
into its subclasses.

Furthermore, the compatible type of an enum is not yet fully determined
while we convert the enum and its enumerators.
So we need to adjust a few pieces of code to cope with this
still missing type.
However, this is not worse than the previous situation,
where a (potentially wrong) dummy type was used instead.
And none of this affects code outside of the frontend.

This is related to #348.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@43250 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c