Rewrite SMTHeapWithByteArray to delegate low-level to another heap
To properly support quantifiers, the SMTHeap low-level handling will
need to be rewritten and enhanced. SMTHeapWithByteArray does not
actually need to know the low-level handling logic. Replacing it with
delegate calls reduces code duplication and allows easier proper
quantifier support. As a bonus, it will now be usable with uninterpreted
functions as well.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/predicate-memory-functions@43257 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c