Revert r15739 / 059b6b76 and r15738 / c46fd77e: Simplify casts
It is not totally obvious that simplifying nested casts is always
correct even in the cases here, due to signed vs. unsigned casts.
As it was done only to make expressions easier to read,
it is safer to remove it.
Note that this affects only symbolic exection.
If we want to have such logic again, it would be better to write it from
scratch in the frontend, such that all analyses could benefit from it.

Furthermore, the method "Types.canHoldAllValues()" is also somewhat
questionable, as it treats pointers like ints, for example,
and has at least the bug that it does not treat enums as numeric.
A better version of this API would for example be to just write it for
the relevant cases like CSimpleType.
And the utility class util.Types is hard to find.
So until we actually need this, lets just remove it.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@43329 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c