Do not remove parts of the CFA with labels during simplification
Fixes #1003

Any simplification has the problem that it changes how the CFA looks
internally compared to what the user expects and what they might have
written down in the specification to match.
In general, we do not guarantee any specific shape of the CFA or the
operations on the edges.

However, matching on labels is quite often used in specifications,
and if the simplification removes such a label,
users have been confused about this.
So lets make a special exception for labels and always keep them.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@43346 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c