Improve detection of loop states.
Simply checking whether the current node is in the loop nodes is not enough,
e.g. a function call inside the loop still counts as being part of the loop
for our purposes.
The new solution is more robust but also fails in edge cases involving
recursion, this is acceptable for now.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/invgen-through-sampling@43404 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c