Massively rewrite AssignmentHandler and ArraySliceExpression
The functionality of ArraySliceExpression is considerably changed: now,
it does not auto-add field modifiers to base if possible. It can store
CRightHandSide now, so even function calls; this simplifies things.

Resolving with union support is now done by first canonicalizing
assignments (forcing trailing field accesses from base to modifiers),
then resolving the base. After that, the base is not manipulated when
converting to progenitor span and simple span assignments.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/predicate-memory-functions@43446 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c