Rename MemoryFunctionHandler to MemoryManipulationFunctionHandler
The handler only handles functions that manipulate memory blocks, i.e.
assign to them in some fashion. It does not handle memcmp and memchr
which only read from memory blocks. The function memcmp is implemented
in CExpressionVisitorWithPointerAliasing.handleCmpFunction together with
strcmp and strncmp, while memchr is currently not implemented at all
(and it is hardly ever used).

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/predicate-memory-functions@43458 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c